Necessary changes to work as a `respect` submodule
Before this change, we'd simply copied the `microservices_demo` folder
from `reboot-dev/respect`'s `amsterdam` branch. However, when
`microservices_demo` is a submodule at a different path than the
original `/microservices_demo`, some changes are necessary. This commit
contains all of the changes necessary to let the `microservices-demo`
repo get submoduled in at `resemble/examples/microservices_demo` in the
`reboot-dev/respect` repo.

We expect that we'll undo most of these changes as we de-Bazelify this
repo, but for now this gets us going with a working
`microservices_demo` in `reboot-dev/respect`'s `main` branch.